requirements.txt
flask
pyjwt
requests
gunicorn